fix: Transaction list paging.
When paging the transaction list, we check to see if the number of
records returned by our query is bigger than the page size of the
returned collection to determine if there is a next page. The problem
was that our query was not returning the extra result we needed for that
check to succeed.

While fixing this bug, we also moved the transaction list query to a
repository-style object.
