+++ src/ledger/mod.rs
@@ -3,3 +3,4 @@ pub mod domain;
 pub mod http;
 pub mod models;
 pub mod queries;
+pub mod services;
