+++ src/repos/mod.rs
@@ -1,4 +1,5 @@
 mod email;
+pub mod transactions;
 mod users;
 
 pub use email::{DynEmailRepo, EmailRepo, EmailVerificationError};
